Выполнено распределение кода по классам в зависимости от функционала
