+++ src/RevitReinforcementCoefficient/Models/RevitRepository.cs
@@ -15,7 +15,8 @@ namespace RevitReinforcementCoefficient.Models {
                 new ElementCategoryFilter(BuiltInCategory.OST_StructuralFraming),
                 new ElementCategoryFilter(BuiltInCategory.OST_Columns),
                 new ElementCategoryFilter(BuiltInCategory.OST_StructuralColumns),
-                new ElementCategoryFilter(BuiltInCategory.OST_StructuralFoundation)
+                new ElementCategoryFilter(BuiltInCategory.OST_StructuralFoundation),
+                new ElementCategoryFilter(BuiltInCategory.OST_Rebar)
             });
 
 
@@ -29,14 +30,8 @@ namespace RevitReinforcementCoefficient.Models {
         public Application Application => UIApplication.Application;
         public Document Document => ActiveUIDocument.Document;
 
-        public List<Element> ElementsByFilter => new FilteredElementCollector(Document, Document.ActiveView.Id)
+        public List<Element> ElementsByFilterInActiveView => new FilteredElementCollector(Document, Document.ActiveView.Id)
             .WherePasses(_categoriesFilter)
             .ToList();
-
-        public List<Element> RebarsInActiveView => new FilteredElementCollector(Document, Document.ActiveView.Id)
-            .OfCategory(BuiltInCategory.OST_Rebar)
-            .ToList();
-
-
     }
 }
